W383 near-match wave: 5 MATCHes (frFont x2, fmGslCont, btlEffect func_002C4C80, datCalendar func_00186A40); mdlFile FUN_0031FDE0 channel-mapping bug fixed nd7->4; bpTexCollectLeaves 13->5; mdlEffect 17->8 and 19->8; btlCamera load-order floor documented

+++ src/Event/Fcl/fclDrawSprite.c
@@ -313,7 +313,7 @@ FUN_0040e710(int param_5, int param_6, float param_1, int param_7, int param_8, 
 
   char *pcVar3;
 
-  u32 uVar4;
+  u16 uVar4;
 
   
 
@@ -387,7 +387,7 @@ FUN_0040e930(int param_5,int param_6,float param_1,int param_7,int param_8,int p
 
   char *pcVar3;
 
-  u32 uVar4;
+  u16 uVar4;
